requirements.txt
# Core dependencies
openvino-genai>=2025.1.0
openvino>=2025.1.0
openvino-tokenizers>=2025.1.0

# HuggingFace for model download
huggingface-hub>=0.20.0

# Environment configuration
python-dotenv>=1.0.0

# Benchmark (optional)
psutil>=5.9.0

# Web search (optional, for /web command)
ddgs>=9.0.0

# Multi-line input support
prompt_toolkit>=3.0.0

# Auto-translate search queries to English
deep-translator>=1.11.0
